+++ db/migrate/20220617014130_create_books_providers.rb
@@ -3,7 +3,7 @@ class CreateBooksProviders < ActiveRecord::Migration[6.0]
     create_table :books_providers do |t|
       t.references :provider, null: false, foreign_key: true
       t.references :book, null: false, foreign_key: true
-      t.integer :quantity, null: false
+      t.bigint :quantity, null: false
       t.boolean :hand_flg, null: false, default: false
 
       t.timestamps
