+++ db/migrate/20220617014950_create_details.rb
@@ -2,8 +2,8 @@ class CreateDetails < ActiveRecord::Migration[6.0]
   def change
     create_table :details do |t|
       t.string :name, null: false
-      t.integer :quantity
-      t.integer :price
+      t.bigint :quantity
+      t.bigint :price
       t.text :remark
       t.references :order, null: false, foreign_key: true
